requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
pillow==10.1.0
pyyaml==6.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
